% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImpReg.R
\name{ImpReg}
\alias{ImpReg}
\title{Takes a numeric vector with missing values and outputs a regression model that can be used to
impute those missing values}
\usage{
ImpReg(data, numvar, preds = c("empty"))
}
\arguments{
\item{data}{a dataset}

\item{numvar}{the numeric field in the dataset to have its missing values imputed}

\item{preds}{an optional input; the names of fields to be used as predictors if method = "reg" (regression)}
}
\value{
a model
}
\description{
Takes a numeric vector with missing values and outputs a regression model that can be used to
impute those missing values
}
\examples{
mock_data <- data.frame(pred1 = rep(1:4,10),pred2=rep(1:5,8),missvar=as.numeric(c("NA",rep(1:2,19),"NA")))
ImpReg(mock_data,"missvar",c("pred1","pred2"))
}
