% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mode.R
\name{Mode}
\alias{Mode}
\title{Calculates a mode of a numeric vector}
\usage{
Mode(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a number
}
\description{
Calculates a mode of a numeric vector
}
\examples{
numvec<-as.numeric(c(1,2,3,3,4,4,"NA"))
Mode(numvec)
Note this returns the first mode found
}
