% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modes.R
\name{Modes}
\alias{Modes}
\title{Calculates all modes of a numeric vector}
\usage{
Modes(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a vector
}
\description{
Calculates all modes of a numeric vector
}
\examples{
numvec<-as.numeric(c(1,2,3,3,4,4,"NA"))
Modes(numvec)
Note this returns all modes
}
